fixed ordersSaga workers (resolve doesn't return anything from now), modified a structure of pages
