+++ src/pages/createOrder/GeneralInformationWrapper.tsx
@@ -1,27 +1,21 @@
 import React from "react";
-import { useNavigate, useParams } from "react-router";
+import { useNavigate } from "react-router";
 import { Formik, Form } from "formik";
 import * as Yup from "yup";
 
-import { StepGeneralInformationValues } from "../../../../types";
-import { orderFormContext } from "../../../../context";
+import { StepGeneralInformationValues } from "../../types";
+import { orderFormContext } from "../../context";
 import GeneralInformation from "./steps/generalInformation/GeneralInformation";
 
 const GeneralInformationWrapper: React.FC = () => {
 	const {
-		formState: { generalInformation, status },
+		formState: { generalInformation },
 		updateGeneralInformation,
 		isDocumentsRequired,
 		addStepDocuments,
 		removeStepDocuments,
-		clearContextData,
 	} = React.useContext(orderFormContext);
-	const { orderId } = useParams<"orderId">();
 	const navigate = useNavigate();
-
-	if (!orderId) {
-		return <p>Не знайдено замовлення з таким id</p>;
-	}
 
 	const validationSchema = Yup.object().shape({
 		country: Yup.mixed().required("Значення не повинно бути пустим."),
@@ -48,23 +42,14 @@ const GeneralInformationWrapper: React.FC = () => {
 				),
 			})
 		),
-		trackNumber: Yup.string()
-			.required("Значення не повинно бути пустим.")
-			.min(6, "Значення занадто коротке. Повинно бути 6 символів або більше."),
+		trackNumber: Yup.string().optional().min(6, "Значення занадто коротке. Повинно бути 6 символів або більше."),
 	});
 
 	const submitStep = (data: StepGeneralInformationValues) => {
 		updateGeneralInformation(data);
 
-		const nextStepUrl = isDocumentsRequired
-			? `/user/orders/${orderId}/documents`
-			: `/user/orders/${orderId}/address`;
+		const nextStepUrl = isDocumentsRequired ? "/new-order/documents" : "/new-order/address";
 		navigate(nextStepUrl);
-	};
-
-	const cancelEditing = () => {
-		clearContextData();
-		navigate("/");
 	};
 
 	return (
@@ -83,8 +68,6 @@ const GeneralInformationWrapper: React.FC = () => {
 							resetShopValue={() => setFieldValue("shop", null)}
 							addStepDocuments={addStepDocuments}
 							removeStepDocuments={removeStepDocuments}
-							cancelEditing={cancelEditing}
-							orderStatus={status}
 						/>
 					</Form>
 				)}
