+++ src/pages/createOrder/steps/address/Address.tsx
@@ -1,51 +1,35 @@
 import React from "react";
-import { useNavigate, useParams } from "react-router";
+import { useNavigate } from "react-router";
 import { Formik, Form } from "formik";
 import * as Yup from "yup";
 import { isValidPhoneNumber } from "libphonenumber-js";
 
-import { StepAddressValues, StepperBarItem, OrderProgressStatusLabel } from "../../../../../../types";
-import { orderFormContext } from "../../../../../../context";
-import StepperBar from "../../../../../../components/stepperBar/StepperBar";
-import Input from "../../../../../../components/input/Input";
-import InputPhone from "../../../../../../components/inputPhone/InputPhone";
-import NavigationLink from "../../../../../../components/navigationLink/NavigationLink";
-import NavigationButton from "../../../../../../components/navigationButton/NavigationButton";
-import Button from "../../../../../../components/button/Button";
+import { StepAddressValues, StepperBarItem } from "../../../../types";
+import { orderFormContext } from "../../../../context";
+import StepperBar from "../../../../components/stepperBar/StepperBar";
+import Input from "../../../../components/input/Input";
+import InputPhone from "../../../../components/inputPhone/InputPhone";
+import NavigationLink from "../../../../components/navigationLink/NavigationLink";
+import NavigationButton from "../../../../components/navigationButton/NavigationButton";
+import Button from "../../../../components/button/Button";
 
 import "./address.scss";
 
 const Address: React.FC = () => {
 	const {
-		formState: { address, status },
+		formState: { address },
 		updateAddress,
 		isDocumentsRequired,
-		clearContextData,
 	} = React.useContext(orderFormContext);
 
-	const { orderId } = useParams<"orderId">();
 	const navigate = useNavigate();
 
-	if (!orderId) {
-		return <p>Не знайдено замовлення з таким id</p>;
-	}
-
-	const prevStep = isDocumentsRequired
-		? `/user/orders/${orderId}/documents`
-		: `/user/orders/${orderId}/general-information`;
-	const nextStep = `/user/orders/${orderId}/confirm-data`;
-
-	const shouldDisableAdressInput =
-		status.label === OrderProgressStatusLabel.TRANSIT || status.label === OrderProgressStatusLabel.DEPARTMENT;
+	const prevStep = isDocumentsRequired ? "/new-order/documents" : "/new-order/general-information";
+	const nextStep = "/new-order/confirm-data";
 
 	const submitStep = (data: StepAddressValues) => {
 		updateAddress(data);
 		navigate(nextStep);
-	};
-
-	const cancelEditing = () => {
-		clearContextData();
-		navigate("/");
 	};
 
 	const validationSchema = Yup.object().shape({
@@ -62,20 +46,12 @@ const Address: React.FC = () => {
 
 	const steps: StepperBarItem[] = isDocumentsRequired
 		? [
-				{
-					title: "Інформація про відправлення",
-					status: "completed",
-					url: `/user/orders/${orderId}/general-information`,
-				},
-				{ title: "Документи", status: "completed", url: `/user/orders/${orderId}/documents` },
+				{ title: "Інформація про відправлення", status: "completed", url: "/new-order/general-information" },
+				{ title: "Документи", status: "completed", url: "/new-order/documents" },
 				{ title: "Адреса отримання", status: "editing" },
 		  ]
 		: [
-				{
-					title: "Інформація про відправлення",
-					status: "completed",
-					url: `/user/orders/${orderId}/general-information`,
-				},
+				{ title: "Інформація про відправлення", status: "completed", url: "/new-order/general-information" },
 				{ title: "Адреса отримання", status: "editing" },
 		  ];
 
@@ -89,12 +65,7 @@ const Address: React.FC = () => {
 					<Form className="address-form">
 						<div className="address-form__row">
 							<div className="address-form__column">
-								<Input
-									name="deliveryAddress"
-									id="input_delivery-address"
-									label="Адреса доставки"
-									readOnly={shouldDisableAdressInput}
-								/>
+								<Input name="deliveryAddress" id="input_delivery-address" label="Адреса доставки" />
 							</div>
 							<div className="address-form__column">
 								<InputPhone
@@ -106,7 +77,7 @@ const Address: React.FC = () => {
 							</div>
 						</div>
 						<div className="address-form__row address-form__row--controls">
-							<Button title="Скасувати" type="button" onClick={cancelEditing} />
+							<Button title="Скасувати" type="button" onClick={() => navigate("/")} />
 							<div className="address-form__navigation">
 								<NavigationLink title="Назад" to={prevStep} />
 								<NavigationButton title="Продовжити" iconPosition="right" type="submit" />
