+++ src/pages/createOrder/steps/generalInformation/generalInformation.scss
