+++ src/pages/editOrder/user/steps/generalInformation/generalInformation.scss
